+++ test/Experimental/RAGTools.jl/runtests.jl
@@ -1,0 +1,10 @@
+using Test
+using SparseArrays, LinearAlgebra
+using PromptingTools.Experimental.RAGTools
+
+include("utils.jl")
+include("types.jl")
+include("preparation.jl")
+include("retrieval.jl")
+# include("generation.jl")
+# include("evaluation.jl")